refactor: change to factory pattern, and add luxos discovery
