+++ src/miners/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod commands;
+pub mod commands;
 pub mod factory;
 pub(crate) mod util;
